Moved Maintenance and Dynamics checks to new "ConfigHealthProvider" file Return error alert when submitting form during maintenance
